move get_train_date to model_metadata file
